+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 -   Increase width of dropdown when selecting a curse mod file
 -   Add some basic filtering to not show files for a different loader
+-   Add a warning and button when Fabric API not installed for Fabric loader